pyproject.toml
[project]
name = "bybit-stream"
version = "0.1.0"
description = ""
authors = [
    {name = "www",email = "pyostr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pybit (>=5.12.0,<6.0.0)",
    "redis (>=7.0.1,<8.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
